Trying out new organization.  Break nontrivial handlers out into their own files.
